Using Pandas for transformation/requirement.txt
For the pandas transformation we are going to follow the following steps.
    ---create a virtual environment named "IMBD_package" - python -m venv IMBD_package
    --- Activate the environment - IMBD_package\Scripts\activate
    --- install pandas - pip install pandas


packages: 

numpy-1.26.4 
pandas-2.2.1
